@@ -1,41 +1,52 @@
 apply plugin: 'com.android.application'
 
 repositories {
-    jcenter()
-    maven  { url "http://repo1.maven.org/maven2" }
     maven { url  "http://dl.bintray.com/dasar/maven" }
+    maven { url "https://s3.amazonaws.com/repo.commonsware.com" }
+    maven { url "https://jitpack.io" }
 }
 
-project.ext.supportLib = "24.1.1"
 
 android {
-    compileSdkVersion 24
-    buildToolsVersion "24.0.1"
+    compileSdkVersion project.sdkVersion
     defaultConfig {
         applicationId "org.horaapps.leafpic"
         minSdkVersion 19
-        targetSdkVersion 24
-        versionName "v0.6-beta-1"
-        versionCode 13
+        targetSdkVersion project.sdkVersion
+        versionName "v0.7-alpha-1"
+        versionCode 15
     }
+
     lintOptions {
         disable 'MissingTranslation'
         disable 'ExtraTranslation'
         abortOnError false
     }
+
+    dexOptions {
+        jumboMode = true
+    }
+
     // This is handled for you by the 2.0+ Gradle Plugin
     aaptOptions {
         additionalParameters "--no-version-vectors"
     }
+
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_6
-        targetCompatibility JavaVersion.VERSION_1_6
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
     }
+
+    signingConfigs {
+        release
+    }
+
     buildTypes {
         release {
-            minifyEnabled false
+            minifyEnabled true
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
             resValue "string", "app_name", "LeafPic"
+            signingConfig signingConfigs.release
         }
 
         debug {
