@@ -10,14 +10,15 @@ import android.graphics.Bitmap;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
+import android.os.Handler;
 import android.provider.Settings;
+import android.support.annotation.CallSuper;
 import android.support.annotation.NonNull;
 import android.support.v4.content.ContextCompat;
 import android.support.v4.view.ViewPager;
 import android.support.v7.app.AlertDialog;
 import android.support.v7.widget.Toolbar;
 import android.util.Log;
-import android.view.Display;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.Surface;
@@ -27,32 +28,51 @@ import android.view.animation.AccelerateInterpolator;
 import android.view.animation.DecelerateInterpolator;
 import android.widget.EditText;
 import android.widget.RelativeLayout;
+import android.widget.TextView;
 import android.widget.Toast;
 
 import com.bumptech.glide.Glide;
 import com.mikepenz.community_material_typeface_library.CommunityMaterial;
 import com.mikepenz.google_material_typeface_library.GoogleMaterial;
+import com.orhanobut.hawk.Hawk;
 import com.yalantis.ucrop.UCrop;
 
 import org.horaapps.leafpic.R;
-import org.horaapps.leafpic.SelectAlbumBottomSheet;
+import org.horaapps.leafpic.SelectAlbumBuilder;
 import org.horaapps.leafpic.activities.base.SharedMediaActivity;
 import org.horaapps.leafpic.adapters.MediaPagerAdapter;
 import org.horaapps.leafpic.animations.DepthPageTransformer;
 import org.horaapps.leafpic.data.Album;
-import org.horaapps.leafpic.data.base.SortingMode;
-import org.horaapps.leafpic.data.base.SortingOrder;
+import org.horaapps.leafpic.data.AlbumSettings;
+import org.horaapps.leafpic.data.HandlingAlbums;
+import org.horaapps.leafpic.data.Media;
+import org.horaapps.leafpic.data.MediaHelper;
+import org.horaapps.leafpic.data.StorageHelper;
+import org.horaapps.leafpic.data.filter.MediaFilter;
+import org.horaapps.leafpic.data.provider.CPHelper;
+import org.horaapps.leafpic.data.sort.MediaComparators;
+import org.horaapps.leafpic.data.sort.SortingMode;
+import org.horaapps.leafpic.data.sort.SortingOrder;
 import org.horaapps.leafpic.fragments.ImageFragment;
 import org.horaapps.leafpic.util.AlertDialogsHelper;
-import org.horaapps.leafpic.util.ColorPalette;
-import org.horaapps.leafpic.util.ContentHelper;
+import org.horaapps.leafpic.util.LegacyCompatFileProvider;
 import org.horaapps.leafpic.util.Measure;
-import org.horaapps.leafpic.util.PreferenceUtil;
-import org.horaapps.leafpic.util.SecurityHelper;
+import org.horaapps.leafpic.util.Security;
 import org.horaapps.leafpic.util.StringUtils;
+import org.horaapps.leafpic.util.file.DeleteException;
+import org.horaapps.leafpic.util.preferences.Prefs;
 import org.horaapps.leafpic.views.HackyViewPager;
+import org.horaapps.liz.ColorPalette;
 
 import java.io.File;
+import java.io.InputStream;
+import java.util.ArrayList;
+import java.util.Collections;
+
+import butterknife.BindView;
+import butterknife.ButterKnife;
+import io.reactivex.android.schedulers.AndroidSchedulers;
+import io.reactivex.schedulers.Schedulers;
 
 /**
  * Created by dnld on 18/02/16.
@@ -60,104 +80,169 @@ import java.io.File;
 @SuppressWarnings("ResourceAsColor")
 public class SingleMediaActivity extends SharedMediaActivity {
 
+    private static final String TAG = SingleMediaActivity.class.getSimpleName();
+
+    private static final int SLIDE_SHOW_INTERVAL = 5000;
     private static final String ISLOCKED_ARG = "isLocked";
-    static final String ACTION_OPEN_ALBUM = "android.intent.action.pagerAlbumMedia";
+    public static final String ACTION_OPEN_ALBUM = "org.horaapps.leafpic.intent.VIEW_ALBUM";
+    public static final String ACTION_OPEN_ALBUM_LAZY = "org.horaapps.leafpic.intent.VIEW_ALBUM_LAZY";
     private static final String ACTION_REVIEW = "com.android.camera.action.REVIEW";
 
-    private HackyViewPager mViewPager;
-    private MediaPagerAdapter adapter;
-    private PreferenceUtil SP;
-    private RelativeLayout ActivityBackground;
-    private SelectAlbumBottomSheet bottomSheetDialogFragment;
-    private SecurityHelper securityObj;
-    private Toolbar toolbar;
+
+    @BindView(R.id.photos_pager)
+    HackyViewPager mViewPager;
+
+    @BindView(R.id.PhotoPager_Layout)
+    RelativeLayout activityBackground;
+    @BindView(R.id.toolbar)
+    Toolbar toolbar;
+
     private boolean fullScreenMode, customUri = false;
+    int position;
+
+    private Album album;
+    private ArrayList<Media> media;
+    private MediaPagerAdapter adapter;
+    private boolean isSlideShowOn = false;
+
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_pager);
+        setContentView(R.layout.activity_single_media);
 
-        SP = PreferenceUtil.getInstance(getApplicationContext());
-        toolbar = (Toolbar) findViewById(R.id.toolbar);
-        mViewPager = (HackyViewPager) findViewById(R.id.photos_pager);
-        securityObj= new SecurityHelper(SingleMediaActivity.this);
+        ButterKnife.bind(this);
 
-        if (savedInstanceState != null)
-            mViewPager.setLocked(savedInstanceState.getBoolean(ISLOCKED_ARG, false));
-        try
-        {
-            Album album;
-            if ((getIntent().getAction().equals(Intent.ACTION_VIEW) || getIntent().getAction().equals(ACTION_REVIEW)) && getIntent().getData() != null) {
+        String action = getIntent().getAction();
 
-                String path = ContentHelper.getMediaPath(getApplicationContext(), getIntent().getData());
+        if (action != null) {
+            switch (action) {
+                case ACTION_OPEN_ALBUM:
+                    loadAlbum(getIntent());
+                    break;
+                case ACTION_OPEN_ALBUM_LAZY:
+                    loadAlbumsLazy(getIntent());
+                    break;
+                default:
+                    loadUri(getIntent().getData());
+                    break;
 
-                File file = null;
-                if (path != null)
-                    file = new File(path);
-
-                if (file != null && file.isFile())
-                    //the image is stored in the storage
-                    album = new Album(getApplicationContext(), file);
-                else {
-                    //try to show with Uri
-                    album = new Album(getApplicationContext(), getIntent().getData());
-                    customUri = true;
-                }
-                getAlbums().addAlbum(0, album);
             }
-            initUI();
-            setupUI();
-        } catch (Exception e) { e.printStackTrace(); }
+        }
+
+        /*if (action != null && action.equals(ACTION_OPEN_ALBUM)) {
+            loadAlbum(getIntent());
+        } else if (getIntent().getData() != null) {
+
+        }*/
+
+        if (savedInstanceState != null) {
+            mViewPager.setLocked(savedInstanceState.getBoolean(ISLOCKED_ARG, false));
+        }
+
+        adapter = new MediaPagerAdapter(getSupportFragmentManager(), media);
+        initUi();
     }
 
-    private void initUI() {
+    private void loadAlbum(Intent intent) {
+        album = intent.getParcelableExtra("album");
+        position = intent.getIntExtra("position", 0);
+        media = intent.getParcelableArrayListExtra("media");
+    }
+
+    private void loadAlbumsLazy(Intent intent) {
+        album = intent.getParcelableExtra("album");
+        //position = intent.getIntExtra("position", 0);
+        Media m = intent.getParcelableExtra("media");
+        media = new ArrayList<>();
+        media.add(m);
+        position = 0;
+
+        ArrayList<Media> list = new ArrayList<>();
+
+        CPHelper.getMedia(getApplicationContext(), album)
+                .subscribeOn(Schedulers.io())
+                .observeOn(AndroidSchedulers.mainThread())
+                .filter(media -> MediaFilter.getFilter(album.filterMode()).accept(media) && !media.equals(m))
+                .subscribe(ma -> {
+                            int i = Collections.binarySearch(
+                                    list, ma, MediaComparators.getComparator(album.settings));
+                            if (i < 0) i = ~i;
+                            list.add(i, ma);
+                        },
+                        throwable -> {
+                            Log.wtf("asd", throwable);
+                        },
+                        () -> {
+                            int i = Collections.binarySearch(
+                                    list, m, MediaComparators.getComparator(album.settings));
+                            if (i < 0) i = ~i;
+
+                            list.add(i, m);
+                            media.clear();
+                            media.addAll(list);
+                            adapter.notifyDataSetChanged();
+                            position = i;
+                            mViewPager.setCurrentItem(position);
+
+                            updatePageTitle(position);
+
+                        });
+    }
+
+    private void loadUri(Uri uri) {
+        album = new Album(uri.toString(), uri.getPath());
+        album.settings = AlbumSettings.getDefaults();
+
+        /*
+        String path = StorageHelper.getMediaPath(getApplicationContext(), getIntent().getData());
+                Album album = null;
+
+                if (path != null) {
+                    album = ContentProviderHelper.getAlbumFromMedia(getApplicationContext(), path);
+                    if (album != null) {
+                        //album.updatePhotos(getApplicationContext());
+                        album.setCurrentMedia(path);
+                    }
+                }
+        */
+
+        try {
+            InputStream inputStream = getContentResolver().openInputStream(uri);
+            if (inputStream != null) inputStream.close();
+        } catch (Exception ex) {
+            boolean showEasterEgg = Prefs.showEasterEgg();
+            ((TextView) findViewById(R.id.nothing_to_show_text_emoji_easter_egg)).setText(R.string.error_occured_open_media);
+            findViewById(R.id.nothing_to_show_placeholder).setVisibility(!showEasterEgg ? View.VISIBLE : View.GONE);
+            findViewById(R.id.ll_emoji_easter_egg).setVisibility(showEasterEgg ? View.VISIBLE : View.GONE);
+        }
+
+        media = new ArrayList<>(Collections.singletonList(new Media(uri)));
+        position = 0;
+        customUri = true;
+    }
+
+
+    private void initUi() {
 
         setSupportActionBar(toolbar);
         toolbar.bringToFront();
-        toolbar.setNavigationIcon(getToolbarIcon(CommunityMaterial.Icon.cmd_arrow_left));
-        toolbar.setNavigationOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                onBackPressed();
-            }
-        });
-        setRecentApp(getString(R.string.app_name));
+        toolbar.setNavigationIcon(getToolbarIcon(GoogleMaterial.Icon.gmd_arrow_back));
+        toolbar.setNavigationOnClickListener(v -> onBackPressed());
+
         setupSystemUI();
 
         getWindow().getDecorView().setOnSystemUiVisibilityChangeListener
-                (new View.OnSystemUiVisibilityChangeListener() {
-                    @Override
-                    public void onSystemUiVisibilityChange(int visibility) {
-                        if ((visibility & View.SYSTEM_UI_FLAG_FULLSCREEN) == 0) showSystemUI();
-                        else hideSystemUI();
-                    }
+                (visibility -> {
+                    if ((visibility & View.SYSTEM_UI_FLAG_FULLSCREEN) == 0) showSystemUI();
+                    else hideSystemUI();
                 });
-        adapter = new MediaPagerAdapter(getSupportFragmentManager(), getAlbum().getMedia());
 
-        adapter.setVideoOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                if (SP.getBoolean("set_internal_player", false)) {
-                    Intent mpdIntent = new Intent(SingleMediaActivity.this, PlayerActivity.class)
-                            .setData(getAlbum().getCurrentMedia().getUri());
-                    startActivity(mpdIntent);
-                } else {
-                    Intent intentOpenWith = new Intent(Intent.ACTION_VIEW);
-                    intentOpenWith.setDataAndType(
-                            getAlbum().getMedia().get(mViewPager.getCurrentItem()).getUri(),
-                            getAlbum().getMedia().get(mViewPager.getCurrentItem()).getMimeType());
-                    startActivity(intentOpenWith);
-                }
-            }
-        });
-
-        getSupportActionBar().setTitle((getAlbum().getCurrentMediaIndex() + 1) + " " + getString(R.string.of) + " " + getAlbum().getMedia().size());
+        updatePageTitle(position);
 
         mViewPager.setAdapter(adapter);
-        mViewPager.setCurrentItem(getAlbum().getCurrentMediaIndex());
+        mViewPager.setCurrentItem(position);
         mViewPager.setPageTransformer(true, new DepthPageTransformer());
-        mViewPager.setOffscreenPageLimit(3);
         mViewPager.addOnPageChangeListener(new ViewPager.OnPageChangeListener() {
             @Override
             public void onPageScrolled(int position, float positionOffset, int positionOffsetPixels) {
@@ -165,9 +250,10 @@ public class SingleMediaActivity extends SharedMediaActivity {
 
             @Override
             public void onPageSelected(int position) {
-                getAlbum().setCurrentPhotoIndex(position);
-                toolbar.setTitle((position + 1) + " " + getString(R.string.of) + " " + getAlbum().getMedia().size());
-                invalidateOptionsMenu();
+                SingleMediaActivity.this.position = position;
+
+                updatePageTitle(position);
+                supportInvalidateOptionsMenu();
             }
 
             @Override
@@ -175,40 +261,56 @@ public class SingleMediaActivity extends SharedMediaActivity {
             }
         });
 
-        Display aa = ((WindowManager) getSystemService(WINDOW_SERVICE)).getDefaultDisplay();
-
-        if (aa.getRotation() == Surface.ROTATION_90) {
+        if (((WindowManager) getSystemService(WINDOW_SERVICE)).getDefaultDisplay().getRotation() == Surface.ROTATION_90) {
             Configuration configuration = new Configuration();
             configuration.orientation = Configuration.ORIENTATION_LANDSCAPE;
             onConfigurationChanged(configuration);
         }
-
     }
 
-    private void setupUI() {
+    Handler handler = new Handler();
+    Runnable slideShowRunnable = new Runnable() {
+        @Override
+        public void run() {
+            try {
+                mViewPager.setCurrentItem((mViewPager.getCurrentItem() + 1) % album.getCount());
+            } catch (Exception e) {
+                e.printStackTrace();
+            } finally {
+                handler.postDelayed(this, SLIDE_SHOW_INTERVAL);
+            }
+        }
+    };
 
+    @CallSuper
+    public void updateUiElements() {
+        super.updateUiElements();
         /**** Theme ****/
-        toolbar = (Toolbar) findViewById(R.id.toolbar);
         toolbar.setBackgroundColor(
-                isApplyThemeOnImgAct()
-                        ? ColorPalette.getTransparentColor (getPrimaryColor(), getTransparency())
+                themeOnSingleImgAct()
+                        ? ColorPalette.getTransparentColor(getPrimaryColor(), 255 - Hawk.get(getString(R.string.preference_transparency), 0))
                         : ColorPalette.getTransparentColor(getDefaultThemeToolbarColor3th(), 175));
 
         toolbar.setPopupTheme(getPopupToolbarStyle());
 
-        ActivityBackground = (RelativeLayout) findViewById(R.id.PhotoPager_Layout);
-        ActivityBackground.setBackgroundColor(getBackgroundColor());
+
+        activityBackground.setBackgroundColor(getBackgroundColor());
 
         setStatusBarColor();
         setNavBarColor();
+        setRecentApp(getString(R.string.app_name));
 
-        securityObj.updateSecuritySetting();
+        //TODO: EMOJI EASTER EGG - THERE'S NOTHING TO SHOW
+        ((TextView) findViewById(R.id.emoji_easter_egg)).setTextColor(getSubTextColor());
+        ((TextView) findViewById(R.id.nothing_to_show_text_emoji_easter_egg)).setTextColor(getSubTextColor());
+
 
         /**** SETTINGS ****/
 
-        if (SP.getBoolean("set_max_luminosity", false))
+        if (Prefs.getToggleValue(getString(R.string.preference_max_brightness), false))
             updateBrightness(1.0F);
         else try {
+            // TODO: 12/4/16 redo
             float brightness = android.provider.Settings.System.getInt(
                     getContentResolver(), android.provider.Settings.System.SCREEN_BRIGHTNESS);
             brightness = brightness == 1.0F ? 255.0F : brightness;
@@ -217,16 +319,14 @@ public class SingleMediaActivity extends SharedMediaActivity {
             e.printStackTrace();
         }
 
-        if (SP.getBoolean("set_picture_orientation", false))
+        if (Prefs.getToggleValue(getString(R.string.preference_auto_rotate), false))
             setRequestedOrientation(ActivityInfo.SCREEN_ORIENTATION_SENSOR);
         else setRequestedOrientation(ActivityInfo.SCREEN_ORIENTATION_USER);
+
     }
 
-
-    @Override
-    public void onResume() {
-        super.onResume();
-        setupUI();
+    private void updatePageTitle(int position) {
+        getSupportActionBar().setTitle(getString(R.string.of, position + 1, adapter.getCount()));
     }
 
     @Override
@@ -240,15 +340,19 @@ public class SingleMediaActivity extends SharedMediaActivity {
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.menu_view_pager, menu);
+        if (isSlideShowOn) {
+            getMenuInflater().inflate(R.menu.menu_view_page_slide_on, menu);
+            menu.findItem(R.id.slide_show).setIcon(getToolbarIcon(CommunityMaterial.Icon.cmd_stop_circle_outline));
+        } else {
+            getMenuInflater().inflate(R.menu.menu_view_pager, menu);
 
-        menu.findItem(R.id.action_delete).setIcon(getToolbarIcon(CommunityMaterial.Icon.cmd_delete));
-        menu.findItem(R.id.action_share).setIcon(getToolbarIcon(GoogleMaterial.Icon.gmd_share));
-        menu.findItem(R.id.action_rotate).setIcon(getToolbarIcon(CommunityMaterial.Icon.cmd_rotate_right));
-        menu.findItem(R.id.rotate_right_90).setIcon(getToolbarIcon(CommunityMaterial.Icon.cmd_rotate_right).color(getIconColor()));
-        menu.findItem(R.id.rotate_left_90).setIcon(getToolbarIcon(CommunityMaterial.Icon.cmd_rotate_left).color(getIconColor()));
-        menu.findItem(R.id.rotate_180).setIcon(getToolbarIcon(CommunityMaterial.Icon.cmd_replay).color(getIconColor()));
-
+            menu.findItem(R.id.action_delete).setIcon(getToolbarIcon(CommunityMaterial.Icon.cmd_delete));
+            menu.findItem(R.id.action_share).setIcon(getToolbarIcon(GoogleMaterial.Icon.gmd_share));
+            menu.findItem(R.id.action_rotate).setIcon(getToolbarIcon(CommunityMaterial.Icon.cmd_rotate_right));
+            menu.findItem(R.id.rotate_right_90).setIcon(getToolbarIcon(CommunityMaterial.Icon.cmd_rotate_right).color(getIconColor()));
+            menu.findItem(R.id.rotate_left_90).setIcon(getToolbarIcon(CommunityMaterial.Icon.cmd_rotate_left).color(getIconColor()));
+            menu.findItem(R.id.rotate_180).setIcon(getToolbarIcon(CommunityMaterial.Icon.cmd_replay).color(getIconColor()));
+        }
         return true;
     }
 
@@ -258,29 +362,30 @@ public class SingleMediaActivity extends SharedMediaActivity {
         RelativeLayout.LayoutParams params = new RelativeLayout.LayoutParams(
                 RelativeLayout.LayoutParams.MATCH_PARENT, RelativeLayout.LayoutParams.WRAP_CONTENT);
         if (newConfig.orientation == Configuration.ORIENTATION_LANDSCAPE)
-            params.setMargins(0,0,Measure.getNavigationBarSize(SingleMediaActivity.this).x,0);
+            params.setMargins(0, 0, Measure.getNavigationBarSize(SingleMediaActivity.this).x, 0);
         else
-            params.setMargins(0,0,0,0);
+            params.setMargins(0, 0, 0, 0);
 
         toolbar.setLayoutParams(params);
     }
 
     @Override
     public boolean onPrepareOptionsMenu(final Menu menu) {
+        if (!isSlideShowOn) {
+            menu.setGroupVisible(R.id.only_photos_options, !getCurrentMedia().isVideo());
 
-        menu.setGroupVisible(R.id.only_photos_options, !getAlbum().getCurrentMedia().isVideo());
-
-        if (customUri) {
-            menu.setGroupVisible(R.id.on_internal_storage, false);
-            menu.setGroupVisible(R.id.only_photos_options, false);
-            menu.findItem(R.id.sort_action).setVisible(false);
+            if (customUri) {
+                menu.setGroupVisible(R.id.on_internal_storage, false);
+                menu.setGroupVisible(R.id.only_photos_options, false);
+                menu.findItem(R.id.sort_action).setVisible(false);
+            }
         }
         return super.onPrepareOptionsMenu(menu);
 
     }
 
     @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
+    public void onActivityResult(int requestCode, int resultCode, Intent data) {
         if (data != null && resultCode == RESULT_OK) {
             switch (requestCode) {
                 case UCrop.REQUEST_CROP:
@@ -289,8 +394,8 @@ public class SingleMediaActivity extends SharedMediaActivity {
                         try {
                             //copyFileToDownloads(imageUri);
                             // TODO: 21/08/16 handle this better
-                            if(ContentHelper.copyFile(getApplicationContext(), new File(imageUri.getPath()), new File(getAlbum().getPath()))) {
-                                //((ImageFragment) adapter.getRegisteredFragment(getAlbum().getCurrentMediaIndex())).displayMedia(true);
+                            if (StorageHelper.copyFile(getApplicationContext(), new File(imageUri.getPath()), new File(this.album.getPath()))) {
+                                //((ImageFragment) adapter.getRegisteredFragment(this.album.getCurrentMediaIndex())).displayMedia(true);
                                 Toast.makeText(this, R.string.new_file_created, Toast.LENGTH_SHORT).show();
                             }
                             //adapter.notifyDataSetChanged();
@@ -301,32 +406,42 @@ public class SingleMediaActivity extends SharedMediaActivity {
                         StringUtils.showToast(getApplicationContext(), "errori random");
                     break;
                 default:
+                    super.onActivityResult(requestCode, resultCode, data);
                     break;
             }
         }
     }
 
 
-    private void displayAlbums(boolean reload) {
-        Intent i = new Intent(SingleMediaActivity.this, MainActivity.class);
-        Bundle b = new Bundle();
-        b.putInt(SplashScreen.CONTENT, SplashScreen.ALBUMS_PREFETCHED);
-        if (!reload) i.putExtras(b);
-        startActivity(i);
+    private void displayAlbums() {
+        startActivity(new Intent(getApplicationContext(), MainActivity.class));
         finish();
     }
 
     private void deleteCurrentMedia() {
-        getAlbum().deleteCurrentMedia(getApplicationContext());
-        if (getAlbum().getMedia().size() == 0) {
-            if (customUri) finish();
-            else {
-                getAlbums().removeCurrentAlbum();
-                displayAlbums(false);
-            }
-        }
-        adapter.notifyDataSetChanged();
-        toolbar.setTitle((mViewPager.getCurrentItem() + 1) + " " + getString(R.string.of) + " " + getAlbum().getMedia().size());
+        Media currentMedia = getCurrentMedia();
+
+        MediaHelper.deleteMedia(getApplicationContext(), currentMedia)
+                .subscribeOn(Schedulers.io())
+                .observeOn(AndroidSchedulers.mainThread())
+                .subscribe(deleted -> {
+                            media.remove(deleted);
+                            if (media.size() == 0) {
+                                displayAlbums();
+                            }
+                        },
+                        err -> {
+                            if (err instanceof DeleteException)
+                                Toast.makeText(this, R.string.delete_error, Toast.LENGTH_SHORT).show();
+                            else
+                                Toast.makeText(this, err.getMessage(), Toast.LENGTH_SHORT).show();
+                        },
+                        () -> {
+                            adapter.notifyDataSetChanged();
+                            updatePageTitle(mViewPager.getCurrentItem());
+                        });
+
+
     }
 
     @Override
@@ -334,92 +449,101 @@ public class SingleMediaActivity extends SharedMediaActivity {
         switch (item.getItemId()) {
 
             case R.id.rotate_180:
-                if (!((ImageFragment) adapter.getRegisteredFragment(getAlbum().getCurrentMediaIndex())).rotatePicture(180)) {
-                    Toast.makeText(this, R.string.coming_soon, Toast.LENGTH_SHORT).show();
-                }
+                ((ImageFragment) adapter.getRegisteredFragment(position)).rotatePicture(180);
                 break;
 
             case R.id.rotate_right_90:
-                if (!((ImageFragment) adapter.getRegisteredFragment(getAlbum().getCurrentMediaIndex())).rotatePicture(90)) {
-                    Toast.makeText(this, R.string.coming_soon, Toast.LENGTH_SHORT).show();
-                }
+                ((ImageFragment) adapter.getRegisteredFragment(position)).rotatePicture(90);
                 break;
 
             case R.id.rotate_left_90:
-                if (!((ImageFragment) adapter.getRegisteredFragment(getAlbum().getCurrentMediaIndex())).rotatePicture(-90)) {
-                    Toast.makeText(this, R.string.coming_soon, Toast.LENGTH_SHORT).show();
-                }
+                ((ImageFragment) adapter.getRegisteredFragment(position)).rotatePicture(-90);
                 break;
 
 
             case R.id.action_copy:
-                bottomSheetDialogFragment = new SelectAlbumBottomSheet();
-                bottomSheetDialogFragment.setTitle(getString(R.string.copy_to));
-                bottomSheetDialogFragment.setSelectAlbumInterface(new SelectAlbumBottomSheet.SelectAlbumInterface() {
-                    @Override
-                    public void folderSelected(String path) {
-                        getAlbum().copyPhoto(getApplicationContext(), getAlbum().getCurrentMedia().getPath(), path);
-                        bottomSheetDialogFragment.dismiss();
-                    }
-                });
-                bottomSheetDialogFragment.show(getSupportFragmentManager(), bottomSheetDialogFragment.getTag());
+                SelectAlbumBuilder.with(getSupportFragmentManager())
+                        .title(getString(R.string.copy_to))
+                        .onFolderSelected(path -> {
 
+                            Media currentMedia = getCurrentMedia();
+                            boolean b = MediaHelper.copyMedia(getApplicationContext(), currentMedia, path);
+                            if (!b)
+                                Toast.makeText(getApplicationContext(), R.string.copy_error, Toast.LENGTH_SHORT).show();
+                        }).show();
                 break;
 
-            case R.id.name_sort_action:
-                getAlbum().setDefaultSortingMode(getApplicationContext(), SortingMode.NAME);
-                getAlbum().sortPhotos();
-                adapter.swapDataSet(getAlbum().getMedia());
+            case R.id.name_sort_mode:
+                HandlingAlbums.getInstance(getApplicationContext())
+                        .setSortingMode(album.getPath(), SortingMode.NAME.getValue());
+                album.setSortingMode(SortingMode.NAME);
+                this.album.sortPhotos();
+                adapter.swapDataSet(media);
                 item.setChecked(true);
                 return true;
 
-            case R.id.date_taken_sort_action:
-                getAlbum().setDefaultSortingMode(getApplicationContext(), SortingMode.DATE);
-                getAlbum().sortPhotos();
-                adapter.swapDataSet(getAlbum().getMedia());
+            case R.id.date_taken_sort_mode:
+                HandlingAlbums.getInstance(getApplicationContext())
+                        .setSortingMode(album.getPath(), SortingMode.DATE.getValue());
+                album.setSortingMode(SortingMode.DATE);
+                this.album.sortPhotos();
+                adapter.swapDataSet(media);
                 item.setChecked(true);
                 return true;
 
-            case R.id.size_sort_action:
-                getAlbum().setDefaultSortingMode(getApplicationContext(), SortingMode.SIZE);
-                getAlbum().sortPhotos();
-                adapter.swapDataSet(getAlbum().getMedia());
+            case R.id.size_sort_mode:
+                HandlingAlbums.getInstance(getApplicationContext())
+                        .setSortingMode(album.getPath(), SortingMode.SIZE.getValue());
+                album.setSortingMode(SortingMode.SIZE);
+                this.album.sortPhotos();
+                adapter.swapDataSet(media);
                 item.setChecked(true);
                 return true;
 
             case R.id.type_sort_action:
-                getAlbum().setDefaultSortingMode(getApplicationContext(), SortingMode.TYPE);
-                getAlbum().sortPhotos();
-                adapter.swapDataSet(getAlbum().getMedia());
+                HandlingAlbums.getInstance(getApplicationContext())
+                        .setSortingMode(album.getPath(), SortingMode.TYPE.getValue());
+                album.setSortingMode(SortingMode.TYPE);
+                this.album.sortPhotos();
+                adapter.swapDataSet(media);
                 item.setChecked(true);
                 return true;
 
-            case R.id.numeric_sort_action:
-                getAlbum().setDefaultSortingMode(getApplicationContext(), SortingMode.NUMERIC);
-                getAlbum().sortPhotos();
-                adapter.swapDataSet(getAlbum().getMedia());
+            case R.id.numeric_sort_mode:
+                HandlingAlbums.getInstance(getApplicationContext())
+                        .setSortingMode(album.getPath(), SortingMode.NUMERIC.getValue());
+                album.setSortingMode(SortingMode.NUMERIC);
+                this.album.sortPhotos();
+                adapter.swapDataSet(media);
                 item.setChecked(true);
                 return true;
 
-            case R.id.ascending_sort_action:
-                getAlbum().setDefaultSortingAscending(getApplicationContext(), !item.isChecked() ? SortingOrder.ASCENDING : SortingOrder.DESCENDING);
-                getAlbum().sortPhotos();
-                adapter.swapDataSet(getAlbum().getMedia());
-
+            case R.id.ascending_sort_order:
                 item.setChecked(!item.isChecked());
+                SortingOrder sortingOrder = SortingOrder.fromValue(item.isChecked());
+
+                HandlingAlbums.getInstance(getApplicationContext())
+                        .setSortingOrder(album.getPath(), sortingOrder.getValue());
+                album.setSortingOrder(sortingOrder);
+                this.album.sortPhotos();
+                adapter.swapDataSet(media);
                 return true;
 
 
             case R.id.action_share:
+                // TODO: 16/10/17 check if it works everywhere
                 Intent share = new Intent(Intent.ACTION_SEND);
-                share.setType(getAlbum().getCurrentMedia().getMimeType());
-                share.putExtra(Intent.EXTRA_STREAM, getAlbum().getCurrentMedia().getUri());
+                share.setType(getCurrentMedia().getMimeType());
+                Uri uri1 = LegacyCompatFileProvider.getUri(this, getCurrentMedia().getFile());
+                share.putExtra(Intent.EXTRA_STREAM, uri1);
+                share.setFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
                 startActivity(Intent.createChooser(share, getString(R.string.send_to)));
                 return true;
 
             case R.id.action_edit:
+                // TODO: 16/10/17 redo
                 Uri mDestinationUri = Uri.fromFile(new File(getCacheDir(), "croppedImage.png"));
-                Uri uri = Uri.fromFile(new File(getAlbum().getCurrentMedia().getPath()));
+                Uri uri = Uri.fromFile(new File(getCurrentMedia().getPath()));
                 UCrop uCrop = UCrop.of(uri, mDestinationUri);
                 uCrop.withOptions(getUcropOptions());
                 uCrop.start(SingleMediaActivity.this);
@@ -427,145 +551,143 @@ public class SingleMediaActivity extends SharedMediaActivity {
 
             case R.id.action_use_as:
                 Intent intent = new Intent(Intent.ACTION_ATTACH_DATA);
-                intent.setDataAndType(
-                        getAlbum().getCurrentMedia().getUri(), getAlbum().getCurrentMedia().getMimeType());
+                intent.setDataAndType(LegacyCompatFileProvider.getUri(this,
+                        getCurrentMedia().getFile()), getCurrentMedia().getMimeType());
+                intent.setFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
                 startActivity(Intent.createChooser(intent, getString(R.string.use_as)));
                 return true;
 
             case R.id.action_open_with:
                 Intent intentopenWith = new Intent(Intent.ACTION_VIEW);
-                intentopenWith.setDataAndType(
-                        getAlbum().getCurrentMedia().getUri(), getAlbum().getCurrentMedia().getMimeType());
+                intentopenWith.setDataAndType(LegacyCompatFileProvider.getUri(this,
+                        getCurrentMedia().getFile()), getCurrentMedia().getMimeType());
+                intentopenWith.setFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
                 startActivity(Intent.createChooser(intentopenWith, getString(R.string.open_with)));
                 break;
 
             case R.id.action_delete:
-                final AlertDialog.Builder deleteDialog = new AlertDialog.Builder(SingleMediaActivity.this, getDialogStyle());
-
-                AlertDialogsHelper.getTextDialog(SingleMediaActivity.this,deleteDialog,
-                        R.string.delete, R.string.delete_photo_message);
-
-                deleteDialog.setNegativeButton(this.getString(R.string.cancel).toUpperCase(), null);
-                deleteDialog.setPositiveButton(this.getString(R.string.delete).toUpperCase(), new DialogInterface.OnClickListener() {
+                final AlertDialog textDialog = AlertDialogsHelper.getTextDialog(SingleMediaActivity.this, R.string.delete, R.string.delete_photo_message);
+                textDialog.setButton(DialogInterface.BUTTON_NEGATIVE, this.getString(R.string.cancel).toUpperCase(), new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialogInterface, int i) {
+                        textDialog.dismiss();
+                    }
+                });
+                textDialog.setButton(DialogInterface.BUTTON_POSITIVE, this.getString(R.string.delete).toUpperCase(), new DialogInterface.OnClickListener() {
                     public void onClick(DialogInterface dialog, int id) {
-                        if (securityObj.isActiveSecurity()&&securityObj.isPasswordOnDelete()) {
+                        if (Security.isPasswordOnDelete()) {
 
-                            final AlertDialog.Builder passwordDialogBuilder = new AlertDialog.Builder(SingleMediaActivity.this, getDialogStyle());
-                            final EditText editTextPassword = securityObj.getInsertPasswordDialog
-                                    (SingleMediaActivity.this, passwordDialogBuilder);
-
-                            passwordDialogBuilder.setPositiveButton(getString(R.string.ok_action).toUpperCase(), new DialogInterface.OnClickListener() {
-                                public void onClick(DialogInterface dialog, int which) {
-                                    if (securityObj.checkPassword(editTextPassword.getText().toString())) {
-                                        deleteCurrentMedia();
-                                    } else
-                                        Toast.makeText(passwordDialogBuilder.getContext(), R.string.wrong_password, Toast.LENGTH_SHORT).show();
-
-                                }
-                            });
-                            passwordDialogBuilder.setNegativeButton(getString(R.string.cancel).toUpperCase(), null);
-                            final AlertDialog passwordDialog = passwordDialogBuilder.create();
-                            passwordDialog.show();
-                            passwordDialog.getButton(AlertDialog.BUTTON_POSITIVE).setOnClickListener(new View
-                                    .OnClickListener() {
+                            Security.authenticateUser(SingleMediaActivity.this, new Security.AuthCallBack() {
                                 @Override
-                                public void onClick(View v) {
-                                    if (securityObj.checkPassword(editTextPassword.getText().toString())){
-                                        deleteCurrentMedia();
-                                        passwordDialog.dismiss();
-                                    } else {
-                                        Toast.makeText(getApplicationContext(), R.string.wrong_password, Toast.LENGTH_SHORT).show();
-                                        editTextPassword.getText().clear();
-                                        editTextPassword.requestFocus();
-                                    }
+                                public void onAuthenticated() {
+                                    deleteCurrentMedia();
+                                }
+
+                                @Override
+                                public void onError() {
+                                    Toast.makeText(getApplicationContext(), R.string.wrong_password, Toast.LENGTH_SHORT).show();
                                 }
                             });
                         } else
                             deleteCurrentMedia();
                     }
                 });
-                deleteDialog.show();
+                textDialog.show();
                 return true;
 
             case R.id.action_move:
-                bottomSheetDialogFragment = new SelectAlbumBottomSheet();
-                bottomSheetDialogFragment.setTitle(getString(R.string.move_to));
-                bottomSheetDialogFragment.setSelectAlbumInterface(new SelectAlbumBottomSheet.SelectAlbumInterface() {
-                    @Override
-                    public void folderSelected(String path) {
-                        getAlbum().moveCurrentMedia(getApplicationContext(), path);
+                SelectAlbumBuilder.with(getSupportFragmentManager())
+                        .title(getString(R.string.move_to))
+                        .exploreMode(true)
+                        .force(true)
+                        .onFolderSelected(path -> {
 
-                        if (getAlbum().getMedia().size() == 0) {
-                            if (customUri) finish();
-                            else {
-                                getAlbums().removeCurrentAlbum();
-                                //((MyApplication) getApplicationContext()).removeCurrentAlbum();
-                                displayAlbums(false);
+                            Media currentMedia = getCurrentMedia();
+
+                            boolean success = MediaHelper.moveMedia(getApplicationContext(), currentMedia, path);
+
+                            if (success) {
+                                media.remove(currentMedia);
+
+                                if (media.size() == 0) {
+                                    displayAlbums();
+                                }
+                            } else {
+                                Toast.makeText(getApplicationContext(), R.string.move_error, Toast.LENGTH_SHORT).show();
                             }
-                        }
-                        adapter.notifyDataSetChanged();
-                        toolbar.setTitle((mViewPager.getCurrentItem() + 1) + " " + getString(R.string.of) + " " + getAlbum().getCount());
-                        bottomSheetDialogFragment.dismiss();
-                    }
-                });
-                bottomSheetDialogFragment.show(getSupportFragmentManager(), bottomSheetDialogFragment.getTag());
+                            adapter.notifyDataSetChanged();
+                            updatePageTitle(mViewPager.getCurrentItem());
+                        }).show();
 
                 return true;
 
             case R.id.action_rename:
-                AlertDialog.Builder renameDialogBuilder = new AlertDialog.Builder(SingleMediaActivity.this, getDialogStyle());
-                final EditText editTextNewName = new EditText(getApplicationContext());
-                editTextNewName.setText(StringUtils.getPhotoNameByPath(getAlbum().getCurrentMedia().getPath()));
+                final EditText editTextNewName = new EditText(this);
+                editTextNewName.setText(StringUtils.getPhotoNameByPath(getCurrentMedia().getPath()));
 
-                AlertDialog renameDialog =
-                        AlertDialogsHelper.getInsertTextDialog(
-                                this,renameDialogBuilder, editTextNewName, R.string.rename_photo_action);
+                AlertDialog renameDialog = AlertDialogsHelper.getInsertTextDialog(this, editTextNewName, R.string.rename_photo_action);
 
                 renameDialog.setButton(DialogInterface.BUTTON_POSITIVE, getString(R.string.ok_action).toUpperCase(), new DialogInterface.OnClickListener() {
                     @Override
                     public void onClick(DialogInterface dialog, int which) {
-                        if (editTextNewName.length() != 0)
-                            getAlbum().renameCurrentMedia(getApplicationContext(), editTextNewName.getText().toString());
-                        else
+                        if (editTextNewName.length() != 0) {
+                            Media currentMedia = getCurrentMedia();
+                            boolean b = MediaHelper.renameMedia(getApplicationContext(), currentMedia, editTextNewName.getText().toString());
+                            if (!b) {
+                                StringUtils.showToast(getApplicationContext(), getString(R.string.rename_error));
+                                //adapter.notifyDataSetChanged();
+                            }
+                        } else
                             StringUtils.showToast(getApplicationContext(), getString(R.string.nothing_changed));
-                    }});
-                renameDialog.setButton(DialogInterface.BUTTON_NEGATIVE, getString(R.string.cancel).toUpperCase(), new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialog, int which) { } });
+                    }
+                });
+                renameDialog.setButton(DialogInterface.BUTTON_NEGATIVE, getString(R.string.cancel).toUpperCase(), (dialog, which) -> dialog.dismiss());
                 renameDialog.show();
                 break;
 
             case R.id.action_edit_with:
                 Intent editIntent = new Intent(Intent.ACTION_EDIT);
-                editIntent.setDataAndType(getAlbum().getCurrentMedia().getUri(), getAlbum().getCurrentMedia().getMimeType());
+                editIntent.setDataAndType(LegacyCompatFileProvider.getUri(this,
+                        getCurrentMedia().getFile()), getCurrentMedia().getMimeType());
                 editIntent.setFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
-                startActivity(Intent.createChooser(editIntent, "Edit with"));
+                startActivity(Intent.createChooser(editIntent, getString(R.string.edit_with)));
                 break;
 
             case R.id.action_details:
-                AlertDialog.Builder detailsDialogBuilder = new AlertDialog.Builder(SingleMediaActivity.this, getDialogStyle());
-                final AlertDialog detailsDialog =
-                        AlertDialogsHelper.getDetailsDialog(this, detailsDialogBuilder,getAlbum().getCurrentMedia());
+
+                final AlertDialog detailsDialog = AlertDialogsHelper.getDetailsDialog(this, getCurrentMedia());
 
                 detailsDialog.setButton(DialogInterface.BUTTON_POSITIVE, getString(R.string
-                        .ok_action).toUpperCase(), new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialog, int which) { }});
+                        .ok_action).toUpperCase(), (dialog, which) -> dialog.dismiss());
 
-                detailsDialog.setButton(DialogInterface.BUTTON_NEUTRAL, getString(R.string.fix_date).toUpperCase(), new DialogInterface.OnClickListener() {
-                    @Override
-                    public void onClick(DialogInterface dialog, int which) {
-                        if (!getAlbum().getCurrentMedia().fixDate())
-                            Toast.makeText(SingleMediaActivity.this, R.string.unable_to_fix_date, Toast.LENGTH_SHORT).show();
-                    }
+                detailsDialog.setButton(DialogInterface.BUTTON_NEUTRAL, getString(R.string.fix_date).toUpperCase(), (dialog, which) -> {
+                    // todo
+                    //if (!getCurrentMedia().fixDate())
+                    Toast.makeText(SingleMediaActivity.this, R.string.unable_to_fix_date, Toast.LENGTH_SHORT).show();
                 });
 
                 detailsDialog.show();
                 break;
 
             case R.id.action_settings:
-                startActivity(new Intent(getApplicationContext(), SettingsActivity.class));
+                SettingsActivity.startActivity(this);
                 break;
+
+            case R.id.action_palette:
+                Intent paletteIntent = new Intent(getApplicationContext(), PaletteActivity.class);
+                paletteIntent.setData(LegacyCompatFileProvider.getUri(this,
+                        getCurrentMedia().getFile()));
+                paletteIntent.setFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION);
+                startActivity(paletteIntent);
+                break;
+
+            case R.id.slide_show:
+                isSlideShowOn = !isSlideShowOn;
+                if (isSlideShowOn) {
+                    handler.postDelayed(slideShowRunnable, SLIDE_SHOW_INTERVAL);
+                    hideSystemUI();
+                } else handler.removeCallbacks(slideShowRunnable);
+                supportInvalidateOptionsMenu();
 
             default:
                 // If we got here, the user's action was not recognized.
@@ -573,6 +695,10 @@ public class SingleMediaActivity extends SharedMediaActivity {
                 //return super.onOptionsItemSelected(item);
         }
         return super.onOptionsItemSelected(item);
+    }
+
+    public Media getCurrentMedia() {
+        return media.get(position);
     }
 
 
@@ -600,7 +726,7 @@ public class SingleMediaActivity extends SharedMediaActivity {
     @Override
     public void setNavBarColor() {
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-            if (isApplyThemeOnImgAct())
+            if (themeOnSingleImgAct())
                 if (isNavigationBarColored())
                     getWindow().setNavigationBarColor(ColorPalette.getTransparentColor(getPrimaryColor(), getTransparency()));
                 else
@@ -613,7 +739,7 @@ public class SingleMediaActivity extends SharedMediaActivity {
     @Override
     protected void setStatusBarColor() {
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-            if (isApplyThemeOnImgAct())
+            if (themeOnSingleImgAct())
                 if (isTranslucentStatusBar() && isTransparencyZero())
                     getWindow().setStatusBarColor(ColorPalette.getObscuredColor(getPrimaryColor()));
                 else
@@ -643,12 +769,18 @@ public class SingleMediaActivity extends SharedMediaActivity {
             public void run() {
                 toolbar.animate().translationY(-toolbar.getHeight()).setInterpolator(new AccelerateInterpolator())
                         .setDuration(200).start();
+                getWindow().getDecorView().setOnSystemUiVisibilityChangeListener(new View.OnSystemUiVisibilityChangeListener() {
+                    @Override
+                    public void onSystemUiVisibilityChange(int visibility) {
+                        Log.wtf(TAG, "ui changed: " + visibility);
+                    }
+                });
                 getWindow().getDecorView().setSystemUiVisibility(
                         View.SYSTEM_UI_FLAG_LAYOUT_STABLE
                                 | View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION
                                 | View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN
-                                | View.SYSTEM_UI_FLAG_HIDE_NAVIGATION // hide nav bar
-                                | View.SYSTEM_UI_FLAG_FULLSCREEN // hide status bar
+                                | View.SYSTEM_UI_FLAG_HIDE_NAVIGATION
+                                | View.SYSTEM_UI_FLAG_FULLSCREEN
                                 | View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY
                                 | View.SYSTEM_UI_FLAG_IMMERSIVE);
 
@@ -697,10 +829,17 @@ public class SingleMediaActivity extends SharedMediaActivity {
         colorAnimation.addUpdateListener(new ValueAnimator.AnimatorUpdateListener() {
             @Override
             public void onAnimationUpdate(ValueAnimator animator) {
-                ActivityBackground.setBackgroundColor((Integer) animator.getAnimatedValue());
+                activityBackground.setBackgroundColor((Integer) animator.getAnimatedValue());
             }
         });
         colorAnimation.start();
+    }
+
+    @Override
+    protected void onDestroy() {
+        super.onDestroy();
+        handler.removeCallbacks(slideShowRunnable);
+        handler = null;
     }
 }
 
