@@ -4,20 +4,23 @@ import android.content.Context;
 
 import com.bumptech.glide.Glide;
 import com.bumptech.glide.GlideBuilder;
-import com.bumptech.glide.load.DecodeFormat;
-import com.bumptech.glide.load.engine.bitmap_recycle.LruBitmapPool;
-import com.bumptech.glide.load.engine.cache.InternalCacheDiskCacheFactory;
-import com.bumptech.glide.load.engine.cache.LruResourceCache;
-import com.bumptech.glide.load.engine.cache.MemorySizeCalculator;
+import com.bumptech.glide.Registry;
 
 /**
  * Created by dnld on 10/03/16.
  */
+
 public class CustomGlideModule implements com.bumptech.glide.module.GlideModule {
+
+    @Override
+    public void registerComponents(Context context, Glide glide, Registry registry) {
+
+    }
+
     @Override
     public void applyOptions(Context context, GlideBuilder builder) {
         // Apply options to the builder here.
-        builder.setDecodeFormat(DecodeFormat.PREFER_ARGB_8888);
+        /*builder.setDecodeFormat(DecodeFormat.PREFER_ARGB_8888);
 
         MemorySizeCalculator calculator = new MemorySizeCalculator(context);
         int defaultMemoryCacheSize = calculator.getMemoryCacheSize();
@@ -33,11 +36,6 @@ public class CustomGlideModule implements com.bumptech.glide.module.GlideModule 
 
         builder.setDiskCache(
                 new InternalCacheDiskCacheFactory(context, cacheSize100MegaBytes)
-        );
-    }
-
-    @Override
-    public void registerComponents(Context context, Glide glide) {
-        // register ModelLoaders here.
+        );*/
     }
 }
