@@ -1,26 +1,29 @@
 package org.horaapps.leafpic.activities;
 
 import android.Manifest;
+import android.app.job.JobInfo;
+import android.app.job.JobScheduler;
+import android.content.ComponentName;
+import android.content.Context;
 import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.graphics.PorterDuff;
-import android.os.AsyncTask;
 import android.os.Build;
 import android.os.Bundle;
+import android.support.annotation.CallSuper;
 import android.support.annotation.NonNull;
 import android.support.v4.content.ContextCompat;
+import android.util.Log;
 import android.view.View;
 import android.widget.ProgressBar;
-import android.widget.RelativeLayout;
 import android.widget.Toast;
 
+import org.horaapps.leafpic.LookForMediaJob;
 import org.horaapps.leafpic.R;
 import org.horaapps.leafpic.activities.base.SharedMediaActivity;
-import org.horaapps.leafpic.data.Album;
-import org.horaapps.leafpic.util.ColorPalette;
 import org.horaapps.leafpic.util.PermissionUtils;
-import org.horaapps.leafpic.util.PreferenceUtil;
 import org.horaapps.leafpic.util.StringUtils;
+import org.horaapps.liz.ColorPalette;
 
 import java.io.File;
 
@@ -29,73 +32,103 @@ import java.io.File;
  */
 public class SplashScreen extends SharedMediaActivity {
 
-    private final int READ_EXTERNAL_STORAGE_ID = 12;
+    private final String TAG = SplashScreen.class.getSimpleName();
+
+    private final int EXTERNAL_STORAGE_PERMISSIONS = 12;
     private static final int PICK_MEDIA_REQUEST = 44;
 
     final static String CONTENT = "content";
-    final static String PICK_MODE = "pick_mode";
 
-    final static int ALBUMS_PREFETCHED = 23;
-    final static int PHOTOS_PREFETCHED = 2;
-    final static int ALBUMS_BACKUP = 60;
-    private boolean PICK_INTENT = false;
-    public final static String ACTION_OPEN_ALBUM = "com.horaapps.leafpic.OPEN_ALBUM";
+    final static int ALBUMS_PREFETCHED = 2376;
+    final static int PHOTOS_PREFETCHED = 2567;
+    final static int ALBUMS_BACKUP = 1312;
+    private boolean pickMode = false;
+    public final static String ACTION_OPEN_ALBUM = "org.horaapps.leafpic.OPEN_ALBUM";
 
-    //private HandlingAlbums albums;
-    private Album album;
-
-    private PreferenceUtil SP;
+    //private Album tmpAlbum;
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(org.horaapps.leafpic.R.layout.activity_splash);
-        SP = PreferenceUtil.getInstance(getApplicationContext());
 
-        ((ProgressBar) findViewById(R.id.progress_splash)).getIndeterminateDrawable().setColorFilter(getPrimaryColor(), PorterDuff.Mode.SRC_ATOP);
-
-        RelativeLayout RelLay = (RelativeLayout) findViewById(org.horaapps.leafpic.R.id.Splah_Bg);
-        RelLay.setBackgroundColor(getBackgroundColor());
 
         getWindow().getDecorView().setSystemUiVisibility(
                 View.SYSTEM_UI_FLAG_LAYOUT_STABLE
                         | View.SYSTEM_UI_FLAG_LAYOUT_HIDE_NAVIGATION
                         | View.SYSTEM_UI_FLAG_LAYOUT_FULLSCREEN);
+
         setNavBarColor();
         setStatusBarColor();
 
-        if (PermissionUtils.isDeviceInfoGranted(this)) {
+        pickMode = getIntent().getAction().equals(Intent.ACTION_GET_CONTENT) || getIntent().getAction().equals(Intent.ACTION_PICK);
+
+        if (PermissionUtils.isStoragePermissionsGranted(this)) {
+
+
             if (getIntent().getAction().equals(ACTION_OPEN_ALBUM)) {
-                Bundle data = getIntent().getExtras();
-                if (data != null) {
-                    String ab = data.getString("albumPath");
+                    Bundle data = getIntent().getExtras();
+                    if (data != null) {
+                        String ab = data.getString("albumPath");
+                        if (ab != null) {
+                            File dir = new File(ab);
+                            //tmpAlbum = new Album(getApplicationContext(), dir.getAbsolutePath(), data.getInt("albumId", -1), dir.getName(), -1);
+                            // TODO: 4/10/17 handle
+                            start();
+                        }
+                    } else StringUtils.showToast(getApplicationContext(), "Album not found");
+                } else  // default intent
+                    start();
 
-                    if (ab != null) {
-                        File dir = new File(ab);
-                        // TODO: 19/08/16 look for id
-                        album = new Album(getApplicationContext(), dir.getAbsolutePath(), data.getInt("albumId", -1), dir.getName(), -1);
-                        new PrefetchPhotosData().execute();
-                    }
-                } else StringUtils.showToast(getApplicationContext(), "Album not found");
-            } else
-                new PrefetchAlbumsData().execute();
 
-            if (getIntent().getAction().equals(Intent.ACTION_GET_CONTENT) || getIntent().getAction().equals(Intent.ACTION_PICK))
-                PICK_INTENT = true;
 
+        } else
+            PermissionUtils.requestPermissions(this, EXTERNAL_STORAGE_PERMISSIONS, Manifest.permission.WRITE_EXTERNAL_STORAGE, Manifest.permission.READ_EXTERNAL_STORAGE);
+
+        //startLookingForMedia();
+    }
+
+    private void start() {
+        Intent intent = new Intent(SplashScreen.this, MainActivity.class);
+
+        if (pickMode) {
+            intent.putExtra(MainActivity.ARGS_PICK_MODE, pickMode);
+            startActivityForResult(intent, PICK_MEDIA_REQUEST);
         } else {
-            String[] permissions = new String[]{Manifest.permission.READ_EXTERNAL_STORAGE};
-            PermissionUtils.requestPermissions(this, READ_EXTERNAL_STORAGE_ID, permissions);
+            startActivity(intent);
+            finish();
         }
     }
 
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-        if (requestCode == PICK_MEDIA_REQUEST) {
-            if (resultCode == RESULT_OK) {
-                setResult(RESULT_OK, data);
-                finish();
+    private void startLookingForMedia() {
+
+        new Thread(() -> {
+            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP /* TODO  && (has included folders) */) {
+
+                JobInfo job = new JobInfo.Builder(0, new ComponentName(getApplicationContext(), LookForMediaJob.class))
+                        .setPeriodic(1000)
+                        .setRequiresDeviceIdle(true)
+                        .build();
+
+                JobScheduler scheduler = (JobScheduler) getSystemService(Context.JOB_SCHEDULER_SERVICE);
+                if (scheduler.getAllPendingJobs().size() == 0)
+                    Log.wtf(TAG, scheduler.schedule(job) == JobScheduler.RESULT_SUCCESS
+                            ? "LookForMediaJob scheduled successfully!" : "LookForMediaJob scheduled failed!");
+
             }
+        }).start();
+    }
+
+    @Override
+    public void onActivityResult(int requestCode, int resultCode, Intent data) {
+        switch (requestCode) {
+            case PICK_MEDIA_REQUEST:
+                if (resultCode == RESULT_OK) {
+                    setResult(RESULT_OK, data);
+                    finish();
+                }
+                break;
+            default: super.onActivityResult(requestCode, resultCode, data);
         }
     }
 
@@ -118,67 +151,30 @@ public class SplashScreen extends SharedMediaActivity {
     @Override
     public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
         switch (requestCode) {
-            case READ_EXTERNAL_STORAGE_ID:
-                boolean granted = grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED;
-                if (granted)
-                    new PrefetchAlbumsData().execute(SP.getBoolean(getString(org.horaapps.leafpic.R.string.preference_auto_update_media), false));
-                else
-                    Toast.makeText(SplashScreen.this, getString(org.horaapps.leafpic.R.string.storage_permission_denied), Toast.LENGTH_LONG).show();
+            case EXTERNAL_STORAGE_PERMISSIONS:
+                boolean gotPermission = grantResults.length > 0;
+
+                for (int result : grantResults) {
+                    gotPermission &= result == PackageManager.PERMISSION_GRANTED;
+                }
+
+                if (gotPermission) {
+                    start();
+                } else {
+                    Toast.makeText(SplashScreen.this, getString(R.string.storage_permission_denied), Toast.LENGTH_LONG).show();
+                    finish();
+                }
                 break;
             default:
                 super.onRequestPermissionsResult(requestCode, permissions, grantResults);
         }
     }
 
-    private class PrefetchAlbumsData extends AsyncTask<Boolean, Boolean, Boolean> {
-
-        @Override
-        protected Boolean doInBackground(Boolean... arg0) {
-            getAlbums().restoreBackup(getApplicationContext());
-            if(getAlbums().dispAlbums.size() == 0) {
-                getAlbums().loadAlbums(getApplicationContext(), false);
-                return true;
-            }
-            return false;
-        }
-
-        @Override
-        protected void onPostExecute(Boolean result) {
-            super.onPostExecute(result);
-
-            Intent i = new Intent(SplashScreen.this, MainActivity.class);
-            Bundle b = new Bundle();
-            b.putInt(CONTENT, result ? ALBUMS_PREFETCHED : ALBUMS_BACKUP);
-            b.putBoolean(PICK_MODE, PICK_INTENT);
-            i.putExtras(b);
-            if (PICK_INTENT)
-                startActivityForResult(i, PICK_MEDIA_REQUEST);
-            else {
-                startActivity(i);
-                finish();
-            }
-            if(result)
-                getAlbums().saveBackup(getApplicationContext());
-        }
-    }
-
-    private class PrefetchPhotosData extends AsyncTask<Void, Void, Void> {
-        @Override
-        protected Void doInBackground(Void... arg0) {
-            album.updatePhotos(getApplicationContext());
-            return null;
-        }
-
-        @Override
-        protected void onPostExecute(Void result) {
-            super.onPostExecute(result);
-            Intent i = new Intent(SplashScreen.this, MainActivity.class);
-            Bundle b = new Bundle();
-            getAlbums().addAlbum(0, album);
-            b.putInt(CONTENT, PHOTOS_PREFETCHED);
-            i.putExtras(b);
-            startActivity(i);
-            finish();
-        }
+    @CallSuper
+    @Override
+    public void updateUiElements() {
+        super.updateUiElements();
+        ((ProgressBar) findViewById(R.id.progress_splash)).getIndeterminateDrawable().setColorFilter(getPrimaryColor(), PorterDuff.Mode.SRC_ATOP);
+        findViewById(org.horaapps.leafpic.R.id.Splah_Bg).setBackgroundColor(getBackgroundColor());
     }
 }
