@@ -1,378 +1,234 @@
 package org.horaapps.leafpic.data;
 
+import android.content.ContentValues;
 import android.content.Context;
-import android.content.Intent;
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
-import android.graphics.Canvas;
-import android.graphics.Color;
-import android.media.MediaScannerConnection;
-import android.media.ThumbnailUtils;
-import android.provider.MediaStore;
+import android.database.Cursor;
+import android.database.sqlite.SQLiteDatabase;
+import android.database.sqlite.SQLiteOpenHelper;
+import android.support.annotation.NonNull;
 
-import org.horaapps.leafpic.R;
-import org.horaapps.leafpic.activities.SplashScreen;
-import org.horaapps.leafpic.data.base.AlbumsComparators;
-import org.horaapps.leafpic.data.base.SortingMode;
-import org.horaapps.leafpic.data.base.SortingOrder;
-import org.horaapps.leafpic.data.providers.MediaStoreProvider;
-import org.horaapps.leafpic.data.providers.StorageProvider;
-import org.horaapps.leafpic.util.ContentHelper;
-import org.horaapps.leafpic.util.PreferenceUtil;
+import org.horaapps.leafpic.data.sort.SortingMode;
+import org.horaapps.leafpic.data.sort.SortingOrder;
 import org.horaapps.leafpic.util.StringUtils;
 
 import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.ObjectInputStream;
-import java.io.ObjectOutputStream;
-import java.io.OptionalDataException;
-import java.io.StreamCorruptedException;
 import java.util.ArrayList;
-import java.util.Collections;
+import java.util.HashSet;
 
 /**
  * Created by dnld on 27/04/16.
  */
-public class HandlingAlbums {
+public class HandlingAlbums extends SQLiteOpenHelper {
 
-  public final static String TAG = "HandlingAlbums";
-  private static String backupFile = "albums.dat";
+    private static final int DATABASE_VERSION = 12;
+    private static final String DATABASE_NAME = "folders.db";
+    private static final String TABLE_ALBUMS = "folders";
 
-  public ArrayList<Album> dispAlbums;
-  private ArrayList<Album> selectedAlbums;
+    public static final int EXCLUDED = 1;
+    public static final int INCLUDED = 2;
 
-  private PreferenceUtil SP;
+    private static final String ALBUM_PATH = "path";
+    private static final String ALBUM_ID = "id";
+    private static final String ALBUM_PINNED = "pinned";
+    private static final String ALBUM_COVER_PATH = "cover_path";
+    private static final String ALBUM_STATUS = "status";
+    private static final String ALBUM_SORTING_MODE = "sorting_mode";
+    private static final String ALBUM_SORTING_ORDER = "sorting_order";
 
-  private int current = 0;
-  private boolean hidden;
+    private static HandlingAlbums mInstance = null;
 
-  public HandlingAlbums(Context context) {
-    SP = PreferenceUtil.getInstance(context);
-    dispAlbums = new ArrayList<Album>();
-    selectedAlbums = new ArrayList<Album>();
-  }
-
-
-  public void loadAlbums(Context context, boolean hidden) {
-    this.hidden = hidden;
-
-    ArrayList<Album> list = new ArrayList<Album>();
-    if (SP.getBoolean(context.getString(org.horaapps.leafpic.R.string.preference_use_alternative_provider), false)) {
-      StorageProvider p = new StorageProvider(context);
-      list = p.getAlbums(context, hidden);
-    } else {
-      list.addAll(MediaStoreProvider.getAlbums(context, hidden));
+    private HandlingAlbums(Context context) {
+        super(context, DATABASE_NAME, null, DATABASE_VERSION);
     }
-    dispAlbums = list;
-    sortAlbums(context);
-  }
 
-  public void addAlbum(int position, Album album) {
-    dispAlbums.add(position, album);
-    setCurrentAlbum(album);
+    public static HandlingAlbums getInstance(Context context) {
+        if(mInstance == null)
+            mInstance = new HandlingAlbums(context);
+        return mInstance;
+    }
 
-  }
+    @Override public void onCreate(SQLiteDatabase db) {
+        db.execSQL("CREATE TABLE " +
+                TABLE_ALBUMS + "(" +
+                ALBUM_PATH + " TEXT," +
+                ALBUM_ID + " INTEGER," +
+                ALBUM_PINNED + " INTEGER," +
+                ALBUM_COVER_PATH + " TEXT, " +
+                ALBUM_STATUS + " INTEGER, " +
+                ALBUM_SORTING_MODE + " INTEGER, " +
+                ALBUM_SORTING_ORDER + " INTEGER)");
 
-  public void setCurrentAlbum(Album album) {
-    current = dispAlbums.indexOf(album);
-  }
+        db.execSQL(String.format("CREATE UNIQUE INDEX idx_path ON %s (%s)", TABLE_ALBUMS, ALBUM_PATH));
+    }
 
+    @Override public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
+        db.execSQL("DROP TABLE IF EXISTS " + TABLE_ALBUMS);
+        db.execSQL("DROP INDEX IF EXISTS idx_path");
+        onCreate(db);
+    }
 
-  public Album getCurrentAlbum() {
-    return dispAlbums.get(current);
-  }
+    public void excludeAlbum(String path) {
+        SQLiteDatabase db = getWritableDatabase();
+        excludeAlbum(db, Album.withPath(path));
+        db.close();
+    }
 
-  public void saveBackup(final Context context) {
-    if (!hidden) {
-      new Thread(new Runnable() {
-        public void run() {
-          FileOutputStream outStream;
-          try {
-            File f = new File(context.getCacheDir(), backupFile);
-            outStream = new FileOutputStream(f);
-            ObjectOutputStream objectOutStream = new ObjectOutputStream(outStream);
-            objectOutStream.writeObject(dispAlbums);
-            objectOutStream.close();
-          } catch (FileNotFoundException e1) {
-            e1.printStackTrace();
-          } catch (IOException e1) {
-            e1.printStackTrace();
-          }
+    private void excludeAlbum(SQLiteDatabase db, Album album) {
+        changeSatusAlbum(db, album, EXCLUDED);
+        // TODO: 3/26/17 notify
+    }
+
+    public void addFolderToWhiteList(String path) {
+        SQLiteDatabase db = getWritableDatabase();
+        changeSatusAlbum(db, Album.withPath(path), INCLUDED);
+        db.close();
+    }
+
+    public ArrayList<String> getExcludedFolders(Context context) {
+        ArrayList<String>  list = new ArrayList<>();
+        HashSet<File> storageRoots = StorageHelper.getStorageRoots(context);
+        for(File file : storageRoots)
+            // it has a lot of garbage
+            list.add(new File(file.getPath(), "Android").getPath());
+
+        list.addAll(getFolders(EXCLUDED));
+        return list;
+    }
+
+    private void changeSatusAlbum(SQLiteDatabase db, Album album, int status) {
+        ContentValues values = new ContentValues();
+        values.put(ALBUM_STATUS, status);
+        if (exist(db, album.getPath())) {
+            db.update(TABLE_ALBUMS, values, ALBUM_PATH+"=?", new String[]{ album.getPath() });
+        } else {
+            values.put(ALBUM_PATH, album.getPath());
+            values.put(ALBUM_PINNED, 0);
+            values.put(ALBUM_SORTING_MODE, SortingMode.DATE.getValue());
+            values.put(ALBUM_SORTING_ORDER, SortingOrder.DESCENDING.getValue());
+            values.put(ALBUM_ID, album.getId());
+            db.insert(TABLE_ALBUMS, null, values);
         }
-      }).start();
     }
-  }
 
-  public static void addAlbumToBackup(final Context context, final Album album) {
-    new Thread(new Runnable() {
-      public void run() {
+    public static ContentValues getDefaults(String path) {
+        ContentValues values = new ContentValues();
+        values.put(ALBUM_PATH, path);
+        values.put(ALBUM_PINNED, 0);
+        values.put(ALBUM_SORTING_MODE, SortingMode.DATE.getValue());
+        values.put(ALBUM_SORTING_ORDER, SortingOrder.DESCENDING.getValue());
+        values.put(ALBUM_ID, -1);
+        return values;
+    }
+
+
+    public void clearStatusFolder(String path) {
+        SQLiteDatabase db = getWritableDatabase();
+        ContentValues values = new ContentValues();
+        values.put(ALBUM_STATUS, "");
+        if (exist(db, path))
+            db.update(TABLE_ALBUMS, values, ALBUM_PATH+"=?", new String[]{ path });
+        // NOTE: it make no difference if the folder was included
+        //excludedFolders.remove(path);
+        db.close();
+    }
+
+    /**
+     *
+     * @param status 1 for EXCLUDED, 2 for INCLUDED
+     * @return
+     */
+    public ArrayList<String> getFolders(int status) {
+        ArrayList<String> list = new ArrayList<>();
+        SQLiteDatabase db = getReadableDatabase();
+        Cursor cur = db.query(TABLE_ALBUMS, new String[]{ALBUM_PATH}, ALBUM_STATUS + "=?", new String[]{String.valueOf(status)}, null, null, null);
+        if (cur.moveToFirst())
+            do list.add(cur.getString(0)); while (cur.moveToNext());
+        cur.close();
+        db.close();
+        return list;
+    }
+
+    public int getFoldersCount(int status) {
+        int c = 0;
+        SQLiteDatabase db = getReadableDatabase();
+        Cursor cur = db.query(TABLE_ALBUMS, new String[]{"count(*)"}, ALBUM_STATUS + "=?", new String[]{String.valueOf(status)}, null, null, null);
+        if (cur.moveToFirst())
+            c = cur.getInt(0);
+        cur.close();
+        db.close();
+        return c;
+    }
+
+    private static boolean exist(SQLiteDatabase db, String path) {
+        Cursor cur = db.rawQuery(
+                String.format("SELECT EXISTS(SELECT 1 FROM %s WHERE %s=? LIMIT 1)", TABLE_ALBUMS, ALBUM_PATH),
+                new String[]{ path });
+        boolean tracked = cur.moveToFirst() &&  cur.getInt(0) == 1;
+        cur.close();
+        return  tracked;
+    }
+
+    public void setPined(String path, boolean pinned) {
+        ContentValues values = new ContentValues();
+        values.put(ALBUM_PINNED, pinned ? 1 : 0);
+        setValue(path, values);
+    }
+
+    public void setCover(String path, String mediaPath) {
+        ContentValues values = new ContentValues();
+        values.put(ALBUM_COVER_PATH, mediaPath);
+        setValue(path, values);
+    }
+
+    public void setSortingMode(String path, int column) {
+        ContentValues values = new ContentValues();
+        values.put(ALBUM_SORTING_MODE, column);
+        setValue(path, values);
+    }
+
+    public void setSortingOrder(String path, int sortingOrder) {
+        ContentValues values = new ContentValues();
+        values.put(ALBUM_SORTING_ORDER, sortingOrder);
+        setValue(path, values);
+    }
+
+    private void setValue(String path, ContentValues values) {
+        SQLiteDatabase db = this.getWritableDatabase();
+        db.update(TABLE_ALBUMS, values, ALBUM_PATH+"=?", new String[]{ path });
+        db.close();
+    }
+
+    @NonNull public static AlbumSettings getSettings(SQLiteDatabase db, String path) {
+        Cursor cursor = null;
         try {
-          boolean success = false;
-          File f = new File(context.getCacheDir(), backupFile);
-          ObjectInputStream reader = new ObjectInputStream(new FileInputStream(f));
-          Object o = reader.readObject();
-          ArrayList<Album> list = null;
-          if (o != null) {
-            list = (ArrayList<Album>) o;
-            for(int i = 0; i < list.size(); i++) {
-              if (list.get(i).equals(album)) {
-                list.remove(i);
-                list.add(i, album);
-                success = true;
-              }
-            }
-          }
+            if (exist(db, path)) {
+                cursor = db.query(
+                        TABLE_ALBUMS,
+                        StringUtils.asArray(
+                                ALBUM_COVER_PATH,
+                                ALBUM_SORTING_MODE,
+                                ALBUM_SORTING_ORDER,
+                                ALBUM_PINNED),
+                        ALBUM_PATH + "=?",
+                        new String[]{path},
+                        null, null, null);
 
-          if (success) {
-            ObjectOutputStream objectOutStream = new ObjectOutputStream(new FileOutputStream(f));
-            objectOutStream.writeObject(list);
-            objectOutStream.close();
-          }
+                if (cursor.moveToFirst())
+                    return new AlbumSettings(
+                            cursor.getString(0),
+                            cursor.getInt(1),
+                            cursor.getInt(2),
+                            cursor.getInt(3));
+            } else
+                db.insert(
+                        TABLE_ALBUMS,
+                        null,
+                        getDefaults(path));
 
-        } catch (FileNotFoundException e1) {
-          e1.printStackTrace();
-        } catch (IOException e1) {
-          e1.printStackTrace();
-        } catch (ClassNotFoundException e) {
-          e.printStackTrace();
+            return AlbumSettings.getDefaults();
+        } finally {
+            if (cursor != null)
+                cursor.close();
         }
-      }
-    }).start();
-  }
-
-
-  public void restoreBackup(Context context) {
-    FileInputStream inStream;
-    try {
-      File f = new File(context.getCacheDir(), backupFile);
-      inStream = new FileInputStream(f);
-      ObjectInputStream objectInStream = new ObjectInputStream(inStream);
-
-      dispAlbums = (ArrayList<Album>) objectInStream.readObject();
-
-      objectInStream.close();
-    } catch (FileNotFoundException e1) {
-      e1.printStackTrace();
-    } catch (ClassNotFoundException e1) {
-      e1.printStackTrace();
-    } catch (OptionalDataException e1) {
-      e1.printStackTrace();
-    } catch (StreamCorruptedException e1) {
-      e1.printStackTrace();
-    } catch (IOException e1) {
-      e1.printStackTrace();
     }
-  }
-
-  private int toggleSelectAlbum(int index) {
-    if (dispAlbums.get(index) != null) {
-      dispAlbums.get(index).setSelected(!dispAlbums.get(index).isSelected());
-      if (dispAlbums.get(index).isSelected()) selectedAlbums.add(dispAlbums.get(index));
-      else selectedAlbums.remove(dispAlbums.get(index));
-    }
-    return index;
-  }
-
-  public int toggleSelectAlbum(Album album) {
-    return toggleSelectAlbum(dispAlbums.indexOf(album));
-  }
-
-  public Album getAlbum(int index){ return dispAlbums.get(index); }
-
-  public void selectAllAlbums() {
-    for (Album dispAlbum : dispAlbums)
-      if (!dispAlbum.isSelected()) {
-        dispAlbum.setSelected(true);
-        selectedAlbums.add(dispAlbum);
-      }
-  }
-
-  public void removeCurrentAlbum(){ dispAlbums.remove(current); }
-
-  public int getSelectedCount() {
-    return selectedAlbums.size();
-  }
-
-  public void clearSelectedAlbums() {
-    for (Album dispAlbum : dispAlbums)
-      dispAlbum.setSelected(false);
-
-    selectedAlbums.clear();
-  }
-
-  public void installShortcutForSelectedAlbums(Context appCtx) {
-    for (Album selectedAlbum : selectedAlbums) {
-
-      Intent shortcutIntent;
-      shortcutIntent = new Intent(appCtx, SplashScreen.class);
-      shortcutIntent.setAction(SplashScreen.ACTION_OPEN_ALBUM);
-      shortcutIntent.putExtra("albumPath", selectedAlbum.getPath());
-      shortcutIntent.putExtra("albumId", selectedAlbum.getId());
-      shortcutIntent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-      shortcutIntent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
-
-      Intent addIntent = new Intent();
-      addIntent.putExtra(Intent.EXTRA_SHORTCUT_INTENT, shortcutIntent);
-      addIntent.putExtra(Intent.EXTRA_SHORTCUT_NAME, selectedAlbum.getName());
-
-      File image = new File(selectedAlbum.getCoverAlbum().getPath());
-      Bitmap bitmap;
-
-      String mime = StringUtils.getMimeType(image.getAbsolutePath());
-
-      if(mime.startsWith("image")) {
-        bitmap = BitmapFactory.decodeFile(image.getAbsolutePath(), new BitmapFactory.Options());
-      } else if(mime.startsWith("video")) {
-        bitmap = ThumbnailUtils.createVideoThumbnail(selectedAlbum.getCoverAlbum().getPath(),
-                MediaStore.Images.Thumbnails.MINI_KIND);
-      } else return;
-      bitmap = Bitmap.createScaledBitmap(getCroppedBitmap(bitmap), 128, 128, false);
-      addIntent.putExtra(Intent.EXTRA_SHORTCUT_ICON, addWhiteBorder(bitmap, 5));
-
-      addIntent.setAction("com.android.launcher.action.INSTALL_SHORTCUT");
-      appCtx.sendBroadcast(addIntent);
-    }
-  }
-
-  private Bitmap addWhiteBorder(Bitmap bmp, int borderSize) {
-    Bitmap bmpWithBorder = Bitmap.createBitmap(bmp.getWidth() + borderSize * 2, bmp.getHeight() + borderSize * 2, bmp.getConfig());
-    Canvas canvas = new Canvas(bmpWithBorder);
-    canvas.drawColor(Color.WHITE);
-    canvas.drawBitmap(bmp, borderSize, borderSize, null);
-    return bmpWithBorder;
-  }
-
-  private Bitmap getCroppedBitmap(Bitmap srcBmp){
-    Bitmap dstBmp;
-    if (srcBmp.getWidth() >= srcBmp.getHeight()){
-      dstBmp = Bitmap.createBitmap(srcBmp,
-              srcBmp.getWidth()/2 - srcBmp.getHeight()/2, 0,
-              srcBmp.getHeight(), srcBmp.getHeight()
-      );
-    } else {
-      dstBmp = Bitmap.createBitmap(srcBmp, 0,
-              srcBmp.getHeight()/2 - srcBmp.getWidth()/2,
-              srcBmp.getWidth(), srcBmp.getWidth()
-      );
-    }
-    return dstBmp;
-  }
-
-  private void scanFile(Context context, String[] path) {   MediaScannerConnection.scanFile(context, path, null, null); }
-
-  public void hideAlbum(String path, Context context) {
-    File dirName = new File(path);
-    File file = new File(dirName, ".nomedia");
-    if (!file.exists()) {
-      try {
-        FileOutputStream out = new FileOutputStream(file);
-        out.flush();
-        out.close();
-        scanFile(context, new String[]{ file.getAbsolutePath() });
-      } catch (Exception e) {
-        e.printStackTrace();
-      }
-    }
-  }
-  public void hideSelectedAlbums(Context context) {
-    for (Album selectedAlbum : selectedAlbums)
-      hideAlbum(selectedAlbum, context);
-    clearSelectedAlbums();
-  }
-
-  private void hideAlbum(final Album a, Context context) {
-    hideAlbum(a.getPath(), context);
-    dispAlbums.remove(a);
-  }
-
-  public void unHideAlbum(String path, Context context) {
-    File dirName = new File(path);
-    File file = new File(dirName, ".nomedia");
-    if (file.exists()) {
-      if (file.delete())
-        scanFile(context, new String[]{ file.getAbsolutePath() });
-    }
-  }
-  public void unHideSelectedAlbums(Context context) {
-    for (Album selectedAlbum : selectedAlbums)
-      unHideAlbum(selectedAlbum, context);
-    clearSelectedAlbums();
-  }
-
-  private void unHideAlbum(final Album a, Context context) {
-    unHideAlbum(a.getPath(), context);
-    dispAlbums.remove(a);
-  }
-
-  public boolean deleteSelectedAlbums(Context context) {
-    boolean success = true;
-
-    for (Album selectedAlbum : selectedAlbums) {
-      int index = dispAlbums.indexOf(selectedAlbum);
-      if(deleteAlbum(selectedAlbum, context))
-        dispAlbums.remove(index);
-      else success = false;
-    }
-    return success;
-  }
-
-  public boolean deleteAlbum(Album album, Context context) {
-    return ContentHelper.deleteFilesInFolder(context, new File(album.getPath()));
-  }
-
-  public void excludeSelectedAlbums(Context context) {
-    for (Album selectedAlbum : selectedAlbums)
-      excludeAlbum(context, selectedAlbum);
-
-    clearSelectedAlbums();
-  }
-
-  private void excludeAlbum(Context context, Album a) {
-    CustomAlbumsHelper h = CustomAlbumsHelper.getInstance(context);
-    h.excludeAlbum(a.getPath());
-    dispAlbums.remove(a);
-  }
-
-  public SortingMode getSortingMode() {
-    return SortingMode.fromValue(SP.getInt("albums_sorting_mode", SortingMode.DATE.getValue()));
-  }
-
-  public SortingOrder getSortingOrder() {
-    return SortingOrder.fromValue(SP.getInt("albums_sorting_order", SortingOrder.DESCENDING.getValue()));
-  }
-
-  public void setDefaultSortingMode(SortingMode sortingMode) {
-    SP.putInt("albums_sorting_mode", sortingMode.getValue());
-  }
-
-  public void setDefaultSortingAscending(SortingOrder sortingOrder) {
-    SP.putInt("albums_sorting_order", sortingOrder.getValue());
-  }
-
-  public void sortAlbums(final Context context) {
-
-    Album camera = null;
-
-    for(Album album : dispAlbums)
-      if (album.getName().equals("Camera") && dispAlbums.remove(album)) {
-        camera = album;
-        break;
-      }
-
-    Collections.sort(dispAlbums, AlbumsComparators.getComparator(getSortingMode(), getSortingOrder()));
-
-    if (camera != null) {
-      camera.setName(context.getString(R.string.camera));
-      dispAlbums.add(0, camera);
-    }
-  }
-
-  public Album getSelectedAlbum(int index) { return selectedAlbums.get(index); }
-
-  public void loadAlbums(Context applicationContext) {
-    loadAlbums(applicationContext, hidden);
-  }
 }
