@@ -6,13 +6,10 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.koushikdutta.ion.Ion;
-
 import org.horaapps.leafpic.activities.SingleMediaActivity;
 import org.horaapps.leafpic.data.Media;
 
-import uk.co.senab.photoview.PhotoView;
-import uk.co.senab.photoview.PhotoViewAttacher;
+import pl.droidsonroids.gif.GifImageView;
 
 /**
  * Created by dnld on 18/02/16.
@@ -34,29 +31,14 @@ public class GifFragment extends Fragment {
 
     @Override public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        gif =  getArguments().getParcelable("gif");
+        gif = getArguments().getParcelable("gif");
     }
 
 
     @Override public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        PhotoView photoView = new PhotoView(container.getContext());
-
-        Ion.with(getContext())
-                .load(gif.getPath())
-                .intoImageView(photoView);
-
-        photoView.setOnPhotoTapListener(new PhotoViewAttacher.OnPhotoTapListener() {
-            @Override
-            public void onPhotoTap(View view, float x, float y) {
-                ((SingleMediaActivity) getActivity()).toggleSystemUI();
-            }
-
-            @Override
-            public void onOutsidePhotoTap() {
-                ((SingleMediaActivity) getActivity()).toggleSystemUI();
-            }
-        });
-
+        GifImageView photoView = new GifImageView(container.getContext());
+        photoView.setImageURI(gif.getUri());
+        photoView.setOnClickListener(view -> ((SingleMediaActivity) getActivity()).toggleSystemUI());
         return photoView;
     }
 }
