@@ -1,10 +1,15 @@
 package org.horaapps.leafpic.util;
 
 import android.content.Context;
-import android.webkit.MimeTypeMap;
+import android.os.Build;
+import android.text.Html;
+import android.text.Spanned;
 import android.widget.Toast;
 
+import org.horaapps.liz.ColorPalette;
+
 import java.util.Date;
+import java.util.Locale;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -13,14 +18,11 @@ import java.util.regex.Pattern;
  */
 public class StringUtils {
 
-    public static String getMimeType(String path) {
-        String extension = path.substring(path.lastIndexOf('.')+1);
-        return  MimeTypeMap.getSingleton().getMimeTypeFromExtension(extension);
+
+    public static String[] asArray(String ... a) {
+        return a;
     }
 
-    public static String getGenericMIME(String mime) {
-        return mime.split("/")[0] + "/*";
-    }
 
     public static String getPhotoNameByPath(String path) {
         String b[] = path.split("/");
@@ -28,16 +30,29 @@ public class StringUtils {
         return fi.substring(0, fi.lastIndexOf('.'));
     }
 
-    public static String getPhotoPathRenamed(String olderPath, String newName) {
-        String c = "", b[] = olderPath.split("/");
-        for (int x = 0; x < b.length - 1; x++) c += b[x] + "/";
-        c += newName;
-        String name = b[b.length - 1];
-        c += name.substring(name.lastIndexOf('.'));
-        return c;
+    @SuppressWarnings("deprecation")
+    public static Spanned html(String s) {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N)
+            return Html.fromHtml(s, Html.FROM_HTML_MODE_LEGACY);
+        else return Html.fromHtml(s);
     }
 
-    static String incrementFileNameSuffix(String name) {
+    public static String getName(String path) {
+        String b[] = path.split("/");
+        return b[b.length - 1];
+    }
+
+    public static String getPhotoPathRenamed(String olderPath, String newName) {
+        StringBuilder c = new StringBuilder();
+        String b[] = olderPath.split("/");
+        for (int x = 0; x < b.length - 1; x++) c.append(b[x]).append("/");
+        c.append(newName);
+        String name = b[b.length - 1];
+        c.append(name.substring(name.lastIndexOf('.')));
+        return c.toString();
+    }
+
+    public static String incrementFileNameSuffix(String name) {
         StringBuilder builder = new StringBuilder();
 
         int dot = name.lastIndexOf('.');
@@ -85,11 +100,40 @@ public class StringUtils {
         t.show();
     }
 
-     public static String humanReadableByteCount(long bytes, boolean si) {
+    public static String humanReadableByteCount(long bytes, boolean si) {
         int unit = si ? 1000 : 1024;
         if (bytes < unit) return bytes + " B";
         int exp = (int) (Math.log(bytes) / Math.log(unit));
         String pre = (si ? "kMGTPE" : "KMGTPE").charAt(exp - 1) + (si ? "" : "i");
-        return String.format("%.1f %sB", bytes / Math.pow(unit, exp), pre);
+        return String.format(Locale.ENGLISH, "%.1f %sB", bytes / Math.pow(unit, exp), pre);
     }
+
+    public static String b(String content) {
+        return String.format(Locale.ENGLISH, "<b>%s</b>", content);
+    }
+
+    public static String i(String content) {
+        return String.format(Locale.ENGLISH, "<i>%s</i>", content);
+    }
+
+    public static Spanned htmlFormat(String content, int color, boolean bold, boolean italic) {
+        String res = content;
+        if (color != -1) res = color(color, res);
+        if (bold) res = b(res);
+        if (italic) res = i(res);
+        return html(res);
+    }
+
+    public static String color(int color, String content) {
+        return String.format(Locale.ENGLISH,
+                "<font color='%s'>%s</font>",
+                ColorPalette.getHexColor(color), content);
+    }
+
+    public static String color(String color, String content) {
+        return String.format(Locale.ENGLISH,
+                "<font color='%s'>%s</font>",
+                color, content);
+    }
+
 }
