@@ -17,7 +17,7 @@ package org.horaapps.leafpic.util.inapppurchase;
 
 /**
  * Exception thrown when something went wrong with in-app billing.
- * An IabException has an associated IabResult (an error).
+ * An IabException has an associated IabResult (an onError).
  * To get the IAB result that caused this exception to be thrown,
  * call {@link #getResult()}.
  */
@@ -38,6 +38,6 @@ public class IabException extends Exception {
         this(new IabResult(response, message), cause);
     }
 
-    /** Returns the IAB result (error) that this exception signals. */
+    /** Returns the IAB result (onError) that this exception signals. */
     public IabResult getResult() { return mResult; }
 }
