@@ -20,7 +20,7 @@ package org.horaapps.leafpic.util.inapppurchase;
  * A result is composed of a response code (an integer) and possibly a
  * message (String). You can get those by calling
  * {@link #getResponse} and {@link #getMessage()}, respectively. You
- * can also inquire whether a result is a success or a failure by
+ * can also inquire whether a result is a onAuthenticated or a failure by
  * calling {@link #isSuccess()} and {@link #isFailure()}.
  */
 public class IabResult {
